perf: eliminate all function calls from OCP1_ISR hot path
Shadow register set means ISR entry/exit has zero stack save.
But jal+prologue+epilogue on every PLIB call inside the ISR still
costs ~6 cycles each. Replaced with direct SFR writes.

PLIB calls removed from ISR:
  TMR5_Start()                  -> T5CONSET = _T5CON_ON_MASK
  TMR4_PeriodSet(period)        -> PR4 = period
  OCMP1_CompareValueSet()       -> OC1R = period - 5
  OCMP1_CompareSecondaryValueSet() -> OC1RS = period - 3

abs() calls removed from ISR:
  abs(delta[axis])      -> abs_delta[axis] (precomputed in STEPPER_LoadSegment)
  abs(seg->rate_delta)  -> field is now uint32_t (always positive by construction)

Type cleanup:
  rate_delta, decel_rate_delta: int32_t -> uint32_t in MotionSegment
  Removed (int32_t) casts from kinematics.c and motion.c (both operands >= 0)

ISR now contains only: array indexing, integer add/compare, 3 SFR stores.

+++ srcs/motion/kinematics.c
@@ -281,7 +281,7 @@ MotionSegment* KINEMATICS_LinearMove(CoordinatePoint start, CoordinatePoint end,
     
     // Rate delta per step — accel phase
     if(accel_steps > 0) {
-        segment_buffer->rate_delta = (int32_t)((segment_buffer->initial_rate - segment_buffer->nominal_rate) / accel_steps);
+        segment_buffer->rate_delta = (segment_buffer->initial_rate - segment_buffer->nominal_rate) / accel_steps;
     } else {
         segment_buffer->rate_delta = 0;
     }
@@ -289,7 +289,7 @@ MotionSegment* KINEMATICS_LinearMove(CoordinatePoint start, CoordinatePoint end,
     // Rate delta per step — decel phase (asymmetric: exit speed may differ from entry speed)
     uint32_t decel_steps_kin = (uint32_t)max_delta - segment_buffer->decelerate_after;
     if (decel_steps_kin > 0) {
-        segment_buffer->decel_rate_delta = (int32_t)((segment_buffer->final_rate - segment_buffer->nominal_rate) / decel_steps_kin);
+        segment_buffer->decel_rate_delta = (segment_buffer->final_rate - segment_buffer->nominal_rate) / decel_steps_kin;
     } else {
         segment_buffer->decel_rate_delta = 0;
     }
